PORT BUG (candidate): UNISON diverges from the plugin after ONE idle sample
plugin_check.py -- the authoritative cross-check built because the user insisted
gating engine B against the port alone was dangerous -- found a divergence in
src/ on its first run. Verified independently before recording.

Against the plugin binary under Unicorn: pluck POLY, MONO retrig, chorus pad,
delay keys and DCO noise are all BIT-EXACT. UNISON pile (patch 61) fails at
-34.6 dB global and -16.3 dB worst-block. A -16.3 dB block is not a rounding
artefact; it is audible.

Narrowed by varying ONLY the idle frames before note-on: cold is bit-exact, one
single idle frame gives -58.1 dB, 441 frames -33.4 dB, 3000 frames -32.6 dB. The
port matches the plugin on UNISON only when the note lands on the very first
sample.

No existing gate could see it. recall_render_ab proves 57 patches bit-exact
against this same oracle and drives every one of them COLD, so a defect needing
one idle sample is structurally invisible to it -- the same shape as the warm
chorus-arm divergence and the MONO retrigger latch, both also invisible to every
cold gate.

Root cause NOT established and a harness defect is not fully excluded. But POLY
and MONO run the identical harness path with idle prefixes and are bit-exact,
which is evidence against that.

Immediate consequence: no engine B result on a UNISON patch may be believed
until this is triaged, because null_b.py compares engine B against src/, and on
UNISON src/ does not currently match the plugin. Engine B could null perfectly
and be wrong.

+++ engine_b/eb_chorus.c
@@ -171,8 +171,14 @@ void eb_chorus_tick_x(eb_chorus_state *s, const eb_chorus_coef *k,
                                                   - v637) + 1))];
     s->t860 = s->line[(int32_t)(mask & (int32_t)(((int64_t)(uint32_t)s->w
                                                   - v637) + 2))];
-    v638 = (float)((double)(v636 * 16384.0f)
-                   - (double)(int)(v636 * 16384.0f));
+    /* The port does this subtraction in DOUBLE. Here it is done in float.
+     * PROVEN, not argued: tools/engineb/fx_chorus_frac_proof.c enumerates
+     * every float bit pattern in (-1024, 1024) -- 2,298,478,592 of them,
+     * strictly containing the MEASURED reachable range of 72..456 samples --
+     * and the two forms agree bit for bit on all of them, 0 mismatches. On the
+     * ESP32-S3 the double form costs 8 soft-float helper calls per sample
+     * (MEASURED-STATIC, cost.py). */
+    v638 = v636 * 16384.0f - (float)(int)(v636 * 16384.0f);
     s->t864 = v638;
     v639 = v635 + s->c816;
     v640 = s->t856 + ((v638 * s->t860) - (v638 * s->t856));
@@ -185,8 +191,7 @@ void eb_chorus_tick_x(eb_chorus_state *s, const eb_chorus_coef *k,
         s->t876 = s->line[(int32_t)(mask & (int32_t)(((int64_t)(uint32_t)s->w
                                                       - v637r) + 2))];
     }
-    v641 = (float)((double)(v639 * 16384.0f)
-                   - (double)(int)(v639 * 16384.0f));
+    v641 = v639 * 16384.0f - (float)(int)(v639 * 16384.0f);   /* same proof */
     s->t880 = v641;
     v642 = s->t872;
     v643 = s->c560;
